refactor: refactory any functions and name of variable
